Refactored out Facebook API
